refactor code, simplify design
