+++ tester/paddle_to_torch/__init__.py
@@ -1,3 +1,3 @@
-from . import paddle_to_torch
+from . import converter
 
-__all__ = ['paddle_to_torch']
+__all__ = ['converter']
