+++ tester/paddle_to_torch/deprecated/paddle2torch_regular_dict.json
