--- tester/paddle_to_torch/paddle2torch_generic_dict.json
@@ -1,98 +1,0 @@
-{
-    "a": "API started with 'a' should be placed here",
-
-    "paddle.add": {
-        "torch_api": "torch.add",
-        "paddle_torch_args_map": {
-            "x": "input",
-            "y": "other"
-        },
-        "min_input_args": 2,
-        "torch_args": [],
-        "torch_kwargs": {
-            "alpha": 1
-        }
-    },
-
-    "paddle.add_n": {
-        "min_input_args": 1,
-        "composite_steps": [
-            {
-                "torch_api": "torch.stack",
-                "torch_args": ["({inputs}, )"],
-                "torch_kwargs": {
-                    "dim": 0
-                }
-            },
-            {
-                "torch_api": "torch.sum",
-                "torch_args": ["{0}"],
-                "torch_kwargs": {
-                    "dim": 0
-                }
-            }
-        ],
-        "description": "torch.stack(inputs).sum(dim=0)"
-    },
-
-    "b": "API started with 'b' should be placed here",
-
-    "c": "API started with 'c' should be placed here",
-
-    "d": "API started with 'd' should be placed here",
-
-    "e": "API started with 'e' should be placed here",
-
-    "f": "API started with 'f' should be placed here",
-
-    "g": "API started with 'g' should be placed here",
-
-    "h": "API started with 'h' should be placed here",
-
-    "i": "API started with 'i' should be placed here",
-
-    "j": "API started with 'j' should be placed here",
-
-    "k": "API started with 'k' should be placed here",
-
-    "l": "API started with 'l' should be placed here",
-
-    "m": "API started with 'm' should be placed here",
-
-    "n": "API started with 'n' should be placed here",
-
-    "o": "API started with 'o' should be placed here",
-
-    "p": "API started with 'p' should be placed here",
-
-    "q": "API started with 'q' should be placed here",
-
-    "r": "API started with 'r' should be placed here",
-
-    "s": "API started with 's' should be placed here",
-
-    "t": "API started with 't' should be placed here",
-
-    "paddle.Tensor.set_": {
-        "torch_api": "torch.Tensor.set_",
-        "paddle_torch_args_map": {
-            "source": "source",
-            "shape": "size",
-            "stride": "stride",
-            "offset": "storage_offset"
-        }
-    },
-
-    "u": "API started with 'u' should be placed here",
-
-    "v": "API started with 'v' should be placed here",
-
-    "w": "API started with 'w' should be placed here",
-
-    "x": "API started with 'x' should be placed here",
-
-    "y": "API started with 'y' should be placed here",
-
-    "z": "API started with 'z' should be placed here"
-
-}
